Flytt kode som sørger for bergegning av utvidet barnetrygd ut i egen fil
